removed resutls and case from the folder.
removed the physical group from the starting mesh. It was causing Elmer to loose track of the bcs after remeshing. gmsh was remeshing it and in the processsing loosing the physical groups. one way to deal with it is to define the physical group in the geo file. the better way is that to remove physical group definition form the original mesh too. Define the BCs based on the surface tag in the original mesh.